webapp/requirements.txt
Click==7.0
Flask==1.0.2
gevent==1.4.0
google-cloud-pubsub==0.45.0
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.17
numpy==1.17.0
protobuf==3.9.1
six==1.12.0
Werkzeug==0.15.5
